[TRANSFORMATIONS][CPU] Move RoPE to common part (#24615)
- *Transferring RoPE operation and transformations for subsequent reuse
in GPU plugin*
