+++ src/plugins/intel_cpu/src/nodes/rope.h
@@ -5,7 +5,7 @@
 #pragma once
 
 #include "node.h"
-#include "transformations/cpu_opset/common/op/rope.hpp"
+#include "ov_ops/rotary_positional_embeddings.hpp"
 
 namespace ov {
 namespace intel_cpu {
@@ -32,7 +32,7 @@ public:
 private:
     struct Executor {
         virtual void execute(dnnl::stream strm,
-                             const RoPENode::Config& config,
+                             const op::internal::RoPE::Config& config,
                              const std::vector<MemoryPtr>& inputs,
                              const std::vector<MemoryPtr>& outputs) = 0;
         virtual ~Executor() = default;
@@ -45,7 +45,7 @@ private:
     struct RoPEExecutorChatGLM;
     template <typename T>
     struct RoPEExecutorQwen;
-    RoPENode::Config m_config;
+    op::internal::RoPE::Config m_config;
     std::shared_ptr<Executor> m_executor;
 };
